+++ packages/data-components/src/components/prevalent-taxa/prevalent-taxa.ts
@@ -3,6 +3,7 @@ import axios from "axios";
 import { hyperStyled } from "@macrostrat/hyper";
 import styles from "./taxa.module.scss";
 import { Spinner } from "@blueprintjs/core";
+import fetch from "cross-fetch";
 
 const h = hyperStyled(styles);