+++ packages/column-components/CHANGELOG.md
@@ -6,7 +6,8 @@ project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.5.1] - 2026-01-17
 
-Modernized color picker component
+- Modernized color picker component
+- Remove last `findDOMNode` usage for React 19 compatibility
 
 ## [1.5.0] - 2026-01-17